All faces now triangulated + normals calculated via triangles. Hemicubes views working.

+++ include/hemicube.hpp
@@ -6,4 +6,4 @@
 #include "model.hpp"
 
 Matrix formHemicubeMVP(const Vec3f& eye, const Vec3f& dir, const Vec3f& up, int size);
-void calcFormFactorsFromModel(const Model& model, const int faceIdx, std::vector<int>& formFactors, int gridSize);
+void calcFormFactorsFromModel(const Model& model, const int faceIdx, std::vector<float>& formFactors, int gridSize);
